Move all class files involved in the 'NDS File System' window to a new folder called 'filesystem'
