+++ ghidra-extension/src/main/java/ndsware/NdswarePlugin.java
@@ -22,6 +22,7 @@ import ghidra.framework.plugintool.PluginInfo;
 import ghidra.framework.plugintool.PluginTool;
 import ghidra.framework.plugintool.util.PluginStatus;
 import ghidra.program.model.listing.Program;
+import ndsware.filesystem.NdsFileSystemProvider;
 
 /**
  * Provide class-level documentation that describes what this plugin does.
