+++ ghidra-extension/src/main/java/ndsware/filesystem/FileDataPanel.java
@@ -1,4 +1,4 @@
-package ndsware;
+package ndsware.filesystem;
 
 import java.awt.BorderLayout;
 import java.awt.Font;
