+++ ghidra-extension/src/main/java/ndsware/filesystem/FileDataTableModel.java
@@ -1,4 +1,4 @@
-package ndsware;
+package ndsware.filesystem;
 
 import java.util.ArrayList;
 import java.util.List;
