+++ ghidra-extension/src/main/java/ndsware/filesystem/NdsFileSystemProvider.java
@@ -1,4 +1,4 @@
-package ndsware;
+package ndsware.filesystem;
 
 import java.awt.BorderLayout;
 import java.util.ArrayList;
